+++ mayan/apps/document_states/serializers.py
@@ -11,7 +11,7 @@ from mayan.apps.user_management.serializers import UserSerializer
 
 from .models import (
     Workflow, WorkflowInstance, WorkflowInstanceLogEntry, WorkflowState,
-    WorkflowTransition
+    WorkflowTransition, WorkflowTransitionField
 )
 
 
@@ -81,18 +81,62 @@ class WorkflowStateSerializer(serializers.HyperlinkedModelSerializer):
         )
 
 
+class WorkflowTransitionFieldSerializer(serializers.HyperlinkedModelSerializer):
+    url = serializers.SerializerMethodField()
+    workflow_transition_url = serializers.SerializerMethodField()
+
+    class Meta:
+        extra_kwargs = {
+            'url': {
+                'lookup_url_kwarg': 'workflow_transition_field_id',
+                'view_name': 'rest_api:workflowtransitionfield-detail'
+            },
+        }
+        fields = (
+            'field_type', 'name', 'help_text', 'id', 'label', 'required',
+            'url', 'widget', 'widget_kwargs', 'workflow_transition_url'
+        )
+        model = WorkflowTransitionField
+
+    def get_url(self, instance):
+        return reverse(
+            'rest_api:workflowtransitionfield-detail', kwargs={
+                'pk': instance.transition.workflow_id,
+                'workflow_transition_id': instance.transition_id,
+                'workflow_transition_field_id': instance.pk
+            }, request=self.context['request'], format=self.context['format']
+        )
+
+    def get_workflow_transition_url(self, instance):
+        return reverse(
+            'rest_api:workflowtransition-detail', kwargs={
+                'pk': instance.transition.workflow_id,
+                'transition_pk': instance.transition_id,
+            }, request=self.context['request'], format=self.context['format']
+        )
+
+
 class WorkflowTransitionSerializer(serializers.HyperlinkedModelSerializer):
     destination_state = WorkflowStateSerializer()
+    field_list_url = serializers.SerializerMethodField()
     origin_state = WorkflowStateSerializer()
     url = serializers.SerializerMethodField()
     workflow_url = serializers.SerializerMethodField()
 
     class Meta:
         fields = (
-            'destination_state', 'id', 'label', 'origin_state', 'url',
-            'workflow_url',
+            'destination_state', 'field_list_url', 'id', 'label',
+            'origin_state', 'url', 'workflow_url',
         )
         model = WorkflowTransition
+
+    def get_field_list_url(self, instance):
+        return reverse(
+            'rest_api:workflowtransitionfield-list', kwargs={
+                'pk': instance.workflow_id,
+                'workflow_transition_id': instance.pk,
+            }, request=self.context['request'], format=self.context['format']
+        )
 
     def get_url(self, instance):
         return reverse(