+++ mayan/apps/document_states/tests/test_api.py
@@ -19,8 +19,9 @@ from .literals import (
 
 from .mixins import (
     DocumentWorkflowAPIViewTestMixin, WorkflowAPIViewTestMixin,
-    WorkflowStateAPIViewTestMixin, WorkflowTransitionAPIViewTestMixin,
-    WorkflowTestMixin
+    WorkflowStateAPIViewTestMixin, WorkflowTestMixin,
+    WorkflowTransitionAPIViewTestMixin,
+    WorkflowTransitionFieldAPIViewTestMixin, WorkflowTransitionFieldTestMixin
 )
 
 
@@ -700,7 +701,7 @@ class WorkflowStatesAPIViewTestCase(
         )
 
 
-class WorkflowTransitionsAPIViewTestCase(
+class WorkflowTransitionAPIViewTestCase(
     WorkflowTransitionAPIViewTestMixin, DocumentTestMixin, WorkflowTestMixin,
     BaseAPITestCase
 ):
@@ -901,4 +902,173 @@ class WorkflowTransitionsAPIViewTestCase(
         self.assertEqual(
             self.test_workflow_transition.destination_state,
             self.test_workflow_state_1
+        )
+
+
+class WorkflowTransitionFieldAPIViewTestCase(
+    WorkflowTransitionFieldAPIViewTestMixin, DocumentTestMixin,
+    WorkflowTestMixin, WorkflowTransitionFieldTestMixin, BaseAPITestCase
+):
+    auto_upload_test_document = False
+
+    def test_workflow_transition_field_create_view_no_permission(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+
+        transition_field_count = self.test_workflow_transition.fields.count()
+
+        response = self._request_test_workflow_transition_field_create_api_view()
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
+        self.test_workflow_transition.refresh_from_db()
+        self.assertEqual(
+            self.test_workflow_transition.fields.count(),
+            transition_field_count
+        )
+
+    def test_workflow_transition_field_create_view_with_access(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_edit
+        )
+
+        transition_field_count = self.test_workflow_transition.fields.count()
+
+        response = self._request_test_workflow_transition_field_create_api_view()
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        self.test_workflow_transition.refresh_from_db()
+        self.assertEqual(
+            self.test_workflow_transition.fields.count(),
+            transition_field_count + 1
+        )
+
+    def test_workflow_transition_field_delete_view_no_permission(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        transition_field_count = self.test_workflow_transition.fields.count()
+
+        response = self._request_test_workflow_transition_field_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
+        self.assertEqual(
+            self.test_workflow_transition.fields.count(),
+            transition_field_count
+        )
+
+    def test_workflow_transition_field_delete_view_with_access(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_edit
+        )
+
+        transition_field_count = self.test_workflow_transition.fields.count()
+
+        response = self._request_test_workflow_transition_field_delete_api_view()
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        self.assertEqual(
+            self.test_workflow_transition.fields.count(),
+            transition_field_count - 1
+        )
+
+    def test_workflow_transition_field_detail_view_no_permission(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        response = self._request_test_workflow_transition_field_detail_api_view()
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertFalse('results' in response.data)
+
+    def test_workflow_transition_field_detail_view_with_access(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_view
+        )
+
+        response = self._request_test_workflow_transition_field_detail_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(
+            response.data['label'],
+            self.test_workflow_transition_field.label
+        )
+
+    def test_workflow_transition_field_edit_via_patch_view_no_permission(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        transition_field_label = self.test_workflow_transition_field.label
+
+        response = self._request_test_workflow_transition_field_edit_via_patch_api_view()
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
+        self.test_workflow_transition_field.refresh_from_db()
+        self.assertEqual(
+            self.test_workflow_transition_field.label, transition_field_label
+        )
+
+    def test_workflow_transition_field_edit_via_patch_view_with_access(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_edit
+        )
+
+        transition_field_label = self.test_workflow_transition_field.label
+
+        response = self._request_test_workflow_transition_field_edit_via_patch_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        self.test_workflow_transition_field.refresh_from_db()
+        self.assertNotEqual(
+            self.test_workflow_transition_field.label, transition_field_label
+        )
+
+    def test_workflow_transition_field_list_view_no_permission(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        response = self._request_test_workflow_transition_field_list_api_view()
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertFalse('results' in response.data)
+
+    def test_workflow_transition_field_list_view_with_access(self):
+        self._create_test_workflow()
+        self._create_test_workflow_states()
+        self._create_test_workflow_transition()
+        self._create_test_workflow_transition_field()
+
+        self.grant_access(
+            obj=self.test_workflow, permission=permission_workflow_view
+        )
+
+        response = self._request_test_workflow_transition_field_list_api_view()
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(
+            response.data['results'][0]['label'],
+            self.test_workflow_transition_field.label
         )